+++ custom_recognition.go
@@ -60,7 +60,7 @@ func _MaaCustomRecognitionCallbackAgent(
 	image, roi uintptr,
 	transArg uintptr,
 	outBox, outDetail uintptr,
-) uint8 {
+) uintptr {
 	// Here, we are simply passing the uint64 value as a pointer
 	// and will not actually dereference this pointer.
 	id := uint64(transArg)
